fix(ci): restore main architecture and i18n gates
