+++ app/src/main/kotlin/com/poyka/ripdpi/ui/screens/home/HomeScreen.kt
@@ -18,15 +18,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.tooling.preview.Preview
 import com.poyka.ripdpi.R
 import com.poyka.ripdpi.activities.ConnectionState
 import com.poyka.ripdpi.activities.HomeDiagnosticsUiState
-import com.poyka.ripdpi.activities.HomeMode
 import com.poyka.ripdpi.activities.HomeModeCardUiState
 import com.poyka.ripdpi.activities.MainUiState
 import com.poyka.ripdpi.data.ConnectionQualitySnapshot
-import com.poyka.ripdpi.data.Mode
 import com.poyka.ripdpi.permissions.PermissionKind
 import com.poyka.ripdpi.permissions.PermissionRecovery
 import com.poyka.ripdpi.subscription.SubscriptionExpirySummaryUiState
@@ -48,11 +45,8 @@ import com.poyka.ripdpi.ui.navigation.Route
 import com.poyka.ripdpi.ui.testing.RipDpiTestTags
 import com.poyka.ripdpi.ui.testing.ripDpiTestTag
 import com.poyka.ripdpi.ui.theme.RipDpiIcons
-import com.poyka.ripdpi.ui.theme.RipDpiTheme
 import com.poyka.ripdpi.ui.theme.RipDpiThemeTokens
 import com.poyka.ripdpi.ui.theme.resolveDegradationTone
-import kotlinx.collections.immutable.persistentListOf
-import kotlin.time.Duration.Companion.ZERO
 
 @Suppress("LongMethod", "CyclomaticComplexMethod", "LongParameterList")
 @Composable
@@ -345,150 +339,5 @@ private fun HomeDegradationStrip(
         secondaryAction = null,
         tone = tone,
         modifier = Modifier.fillMaxWidth(),
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun HomeScreenAllInactivePreview() {
-    RipDpiTheme(themePreference = "light") {
-        HomeScreen(
-            uiState = MainUiState(modeCards = previewHomeModeCards()),
-            onToggleConnection = {},
-            onOpenDiagnostics = {},
-            onOpenHistory = {},
-            onRepairPermission = {},
-            onOpenVpnPermissionDialog = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun HomeScreenBypassActivePreview() {
-    RipDpiTheme(themePreference = "dark") {
-        HomeScreen(
-            uiState = MainUiState(modeCards = previewHomeModeCards(activeMode = HomeMode.LocalDpiBypass)),
-            onToggleConnection = {},
-            onOpenDiagnostics = {},
-            onOpenHistory = {},
-            onRepairPermission = {},
-            onOpenVpnPermissionDialog = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun HomeScreenVpnActivePreview() {
-    RipDpiTheme(themePreference = "light") {
-        HomeScreen(
-            uiState = MainUiState(modeCards = previewHomeModeCards(activeMode = HomeMode.RemoteVpn)),
-            onToggleConnection = {},
-            onOpenDiagnostics = {},
-            onOpenHistory = {},
-            onRepairPermission = {},
-            onOpenVpnPermissionDialog = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun HomeScreenDiagnosticRunningPreview() {
-    RipDpiTheme(themePreference = "dark") {
-        HomeScreen(
-            uiState = MainUiState(modeCards = previewHomeModeCards(loadingMode = HomeMode.Diagnostic)),
-            onToggleConnection = {},
-            onOpenDiagnostics = {},
-            onOpenHistory = {},
-            onRepairPermission = {},
-            onOpenVpnPermissionDialog = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun HomeScreenErrorPreview() {
-    RipDpiTheme(themePreference = "light") {
-        HomeScreen(
-            uiState =
-                MainUiState(
-                    connectionState = ConnectionState.Error,
-                    errorMessage = "Failed to start VPN",
-                    configuredMode = Mode.Proxy,
-                    proxyIp = "127.0.0.1",
-                    proxyPort = "1080",
-                    connectionDuration = ZERO,
-                ),
-            onToggleConnection = {},
-            onOpenDiagnostics = {},
-            onOpenHistory = {},
-            onRepairPermission = {},
-            onOpenVpnPermissionDialog = {},
-        )
-    }
-}
-
-private fun previewHomeModeCards(
-    activeMode: HomeMode? = null,
-    loadingMode: HomeMode? = null,
-) = persistentListOf(
-    previewHomeModeCard(
-        mode = HomeMode.LocalDpiBypass,
-        activeMode = activeMode,
-        loadingMode = loadingMode,
-    ),
-    previewHomeModeCard(
-        mode = HomeMode.RemoteVpn,
-        activeMode = activeMode,
-        loadingMode = loadingMode,
-    ),
-    previewHomeModeCard(
-        mode = HomeMode.Diagnostic,
-        activeMode = activeMode,
-        loadingMode = loadingMode,
-    ),
-)
-
-private fun previewHomeModeCard(
-    mode: HomeMode,
-    activeMode: HomeMode?,
-    loadingMode: HomeMode?,
-): HomeModeCardUiState {
-    val active = mode == activeMode
-    val loading = mode == loadingMode
-    return HomeModeCardUiState(
-        mode = mode,
-        title =
-            when (mode) {
-                HomeMode.LocalDpiBypass -> "Local bypass"
-                HomeMode.RemoteVpn -> "VPN"
-                HomeMode.Diagnostic -> "Network Diagnostic"
-            },
-        primaryLabel =
-            when (mode) {
-                HomeMode.LocalDpiBypass -> "tlsrec_split_host - AdGuard DoH"
-                HomeMode.RemoteVpn -> "relay.example"
-                HomeMode.Diagnostic -> if (loading) "Stage 2 of 4 - Testing TCP" else "No analysis yet"
-            },
-        secondaryLabel = if (active) "Connected 00:18:42" else null,
-        statusLine =
-            when {
-                loading -> "Running"
-                active -> "Connected 00:18:42"
-                else -> "Inactive"
-            },
-        primaryActionLabel =
-            when {
-                mode == HomeMode.Diagnostic -> "Run Scan"
-                active -> "Disable"
-                else -> "Enable"
-            },
-        configureLabel = "Configure",
-        primaryActionEnabled = !loading,
-        isActive = active,
-        isLoading = loading,
     )
 }
