+++ app/src/main/kotlin/com/poyka/ripdpi/ui/screens/home/HomeScreenPreview.kt
@@ -1,0 +1,157 @@
+package com.poyka.ripdpi.ui.screens.home
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.poyka.ripdpi.activities.ConnectionState
+import com.poyka.ripdpi.activities.HomeMode
+import com.poyka.ripdpi.activities.HomeModeCardUiState
+import com.poyka.ripdpi.activities.MainUiState
+import com.poyka.ripdpi.data.Mode
+import com.poyka.ripdpi.ui.theme.RipDpiTheme
+import kotlinx.collections.immutable.persistentListOf
+import kotlin.time.Duration.Companion.ZERO
+
+@Preview(showBackground = true)
+@Composable
+private fun HomeScreenAllInactivePreview() {
+    RipDpiTheme(themePreference = "light") {
+        HomeScreen(
+            uiState = MainUiState(modeCards = previewHomeModeCards()),
+            onToggleConnection = {},
+            onOpenDiagnostics = {},
+            onOpenHistory = {},
+            onRepairPermission = {},
+            onOpenVpnPermissionDialog = {},
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun HomeScreenBypassActivePreview() {
+    RipDpiTheme(themePreference = "dark") {
+        HomeScreen(
+            uiState = MainUiState(modeCards = previewHomeModeCards(activeMode = HomeMode.LocalDpiBypass)),
+            onToggleConnection = {},
+            onOpenDiagnostics = {},
+            onOpenHistory = {},
+            onRepairPermission = {},
+            onOpenVpnPermissionDialog = {},
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun HomeScreenVpnActivePreview() {
+    RipDpiTheme(themePreference = "light") {
+        HomeScreen(
+            uiState = MainUiState(modeCards = previewHomeModeCards(activeMode = HomeMode.RemoteVpn)),
+            onToggleConnection = {},
+            onOpenDiagnostics = {},
+            onOpenHistory = {},
+            onRepairPermission = {},
+            onOpenVpnPermissionDialog = {},
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun HomeScreenDiagnosticRunningPreview() {
+    RipDpiTheme(themePreference = "dark") {
+        HomeScreen(
+            uiState = MainUiState(modeCards = previewHomeModeCards(loadingMode = HomeMode.Diagnostic)),
+            onToggleConnection = {},
+            onOpenDiagnostics = {},
+            onOpenHistory = {},
+            onRepairPermission = {},
+            onOpenVpnPermissionDialog = {},
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun HomeScreenErrorPreview() {
+    RipDpiTheme(themePreference = "light") {
+        HomeScreen(
+            uiState =
+                MainUiState(
+                    connectionState = ConnectionState.Error,
+                    errorMessage = "Failed to start VPN",
+                    configuredMode = Mode.Proxy,
+                    proxyIp = "127.0.0.1",
+                    proxyPort = "1080",
+                    connectionDuration = ZERO,
+                ),
+            onToggleConnection = {},
+            onOpenDiagnostics = {},
+            onOpenHistory = {},
+            onRepairPermission = {},
+            onOpenVpnPermissionDialog = {},
+        )
+    }
+}
+
+private fun previewHomeModeCards(
+    activeMode: HomeMode? = null,
+    loadingMode: HomeMode? = null,
+) = persistentListOf(
+    previewHomeModeCard(
+        mode = HomeMode.LocalDpiBypass,
+        activeMode = activeMode,
+        loadingMode = loadingMode,
+    ),
+    previewHomeModeCard(
+        mode = HomeMode.RemoteVpn,
+        activeMode = activeMode,
+        loadingMode = loadingMode,
+    ),
+    previewHomeModeCard(
+        mode = HomeMode.Diagnostic,
+        activeMode = activeMode,
+        loadingMode = loadingMode,
+    ),
+)
+
+private fun previewHomeModeCard(
+    mode: HomeMode,
+    activeMode: HomeMode?,
+    loadingMode: HomeMode?,
+): HomeModeCardUiState {
+    val active = mode == activeMode
+    val loading = mode == loadingMode
+    return HomeModeCardUiState(
+        mode = mode,
+        title =
+            when (mode) {
+                HomeMode.LocalDpiBypass -> "Local bypass"
+                HomeMode.RemoteVpn -> "VPN"
+                HomeMode.Diagnostic -> "Network Diagnostic"
+            },
+        primaryLabel =
+            when (mode) {
+                HomeMode.LocalDpiBypass -> "tlsrec_split_host - AdGuard DoH"
+                HomeMode.RemoteVpn -> "relay.example"
+                HomeMode.Diagnostic -> if (loading) "Stage 2 of 4 - Testing TCP" else "No analysis yet"
+            },
+        secondaryLabel = if (active) "Connected 00:18:42" else null,
+        statusLine =
+            when {
+                loading -> "Running"
+                active -> "Connected 00:18:42"
+                else -> "Inactive"
+            },
+        primaryActionLabel =
+            when {
+                mode == HomeMode.Diagnostic -> "Run Scan"
+                active -> "Disable"
+                else -> "Enable"
+            },
+        configureLabel = "Configure",
+        primaryActionEnabled = !loading,
+        isActive = active,
+        isLoading = loading,
+    )
+}
