+++ config/i18n/translatable-keys.txt
@@ -2073,9 +2073,8 @@ app:home_approach_title
 app:home_connection_actuator_action_activate
 app:home_connection_actuator_action_android_managed
 app:home_connection_actuator_action_deactivate
-app:home_connection_actuator_action_wait
 app:home_connection_actuator_direct
-app:home_connection_actuator_secure
+app:home_connection_actuator_relay
 app:home_connection_actuator_state_degraded
 app:home_connection_actuator_state_direct
 app:home_connection_actuator_state_direct_degraded
@@ -3859,9 +3858,11 @@ app:vpn_quality_metric_jitter
 app:vpn_quality_metric_loss
 app:vpn_quality_metric_rtt_p50
 app:vpn_quality_strip_body_critical
+app:vpn_quality_strip_body_nominal
 app:vpn_quality_strip_body_warning
 app:vpn_quality_strip_critical_title
 app:vpn_quality_strip_dismiss
+app:vpn_quality_strip_nominal_title
 app:vpn_quality_strip_reprobe
 app:vpn_quality_strip_since_format
 app:vpn_quality_strip_warning_title
